requirements.txt
torch
torchvision
numpy
diffusers
transformers
huggingface_hub
requests
Pillow

accelerate
matplotlib
tqdm
pandas